requirements.txt
flask_sqlalchemy
python-dotenv
email_validator
Flask
WTForms
Flask-WTF
bcrypt
sqlalchemy
cryptography